+++ client/src/components/table/data-row/types.d.ts
@@ -1,8 +1,10 @@
 import { IRowProps } from 'ka-table/props';
 
-export type DataRowProps = IRowProps & {
+export type DataRowProps = Partial<IRowProps> & {
   /** Key of the first column */
   firstColumnKey: string;
   /** Whether the first column should be sticky (when scrolling horizontally) */
   isFirstColumnSticky: boolean;
+
+  stickyColumnKey: string;
 };